Provide individual CPU usage measurement based on idle time
The various methods to determine CPU usage and load have a number of
disadvantages (see also Documentation/cpu-load.txt), and a
straight-forward method to gain usage information about a particular CPU
is lacking. However, in the context of setting CPU affinity and
isolation, it is often required to monitor the effective usage ratio of
a CPU.

This patch adds an additional CPU usage measuring method that is based
on idle time processing. The data are available for every CPU in
/proc/cpuload/cpuN/data in the format "<idletime> <runtime>". The
counters can be reset by writing to /proc/cpuload/resetall for all CPUs
and to /proc/cpuload/cpuN/reset for a particular CPU, respectively.

To calculate the CPU usage since the most recent reset, divide the
runtime by the sum of runtime plus idletime, e.g.

# for i in `ls -1d /proc/cpuload/cpu* | sort -nk1.18`
> do
>   echo "$i: `awk '{ print (100.0*$2) / ($1+$2)"%" }' <$i/data`"
>   echo 1 >$i/reset
> done
/proc/cpuload/cpu0: 72.0048%
/proc/cpuload/cpu1: 5.49522%
/proc/cpuload/cpu2: 0.27916%
/proc/cpuload/cpu3: 32.3493%

Signed-off-by: Carsten Emde <C.Emde@osadl.org>